[FIX] 리턴값 변경

+++ src/main/java/kr/hs/entrydsm/main/service/post/PostDetailService.java
@@ -3,5 +3,7 @@ package kr.hs.entrydsm.main.service.post;
 import kr.hs.entrydsm.main.payload.response.PostDetailResponse;
 
 public interface PostDetailService {
-  PostDetailResponse getDetailPost(Integer id);
+
+  PostDetailResponse getDetailPost(long id);
+
 }
